Correct ua_ntohs.

+++ usr.sbin/ppp/ua.h
@@ -23,7 +23,7 @@
  * OUT OF THE USE OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF
  * SUCH DAMAGE.
  *
- *	$Id:$
+ *	$Id: ua.h,v 1.1 1998/09/04 18:26:00 brian Exp $
  */
 
 #ifdef __i386__		/* Do any other archs not care about alignment ? */
@@ -65,9 +65,9 @@
 
 #  define ua_ntohs(src, tgt)				\
     do {						\
-      u_int32_t __nh;					\
-      __nh = ((u_int32_t)*(u_char *)(src) << 8) |	\
-          (u_int32_t)*((u_char *)(src) + 1);		\
+      u_int16_t __nh;					\
+      __nh = ((u_int16_t)*(u_char *)(src) << 8) |	\
+          (u_int16_t)*((u_char *)(src) + 1);		\
       memcpy((tgt), &__nh, sizeof __nh);		\
     } while (0)
